Add dual RX fallback (primary/secondary receiver with automatic failover)
Adds a second receiver that runs alongside the primary. Both links stay
active for telemetry, but only one drives RC control at a time. When the
primary link enters failsafe, control passes to the secondary; the main
failsafe now only triggers when both links are down.

- Gated by CLI `dual_rx_enabled` (default OFF); new secondary PG_RX_CONFIG
  settings (receiver_type_secondary, serialrx_provider_secondary,
  serialrx_inverted_secondary, serialrx_halfduplex_secondary,
  sbus_sync_interval_secondary).
- New serial function FUNCTION_RX_SERIAL_SECONDARY reuses the free
  FUNCTION_UNUSED_1 bit (1 << 13) so identical protocols can bind to
  distinct UARTs. All serial RX drivers take a port-function argument.
- rxLink_e runtime abstraction with per-link state each cycle and the
  rxGetActiveLink/rxSetActiveLink/rxIsLinkReceivingSignal/rxIsPrimaryFailsafe
  APIs; LOGIC_CONDITION_RX_SELECT_OVERRIDE to force a link.
- OSD OSD_ACTIVE_RX_LINK element (RX:P / RX:S, blinks on secondary/no-link)
  and a "PRI RX FS" multifunction warning. Widened the multifunction warning
  bitfield to 16 bits to fit the added warning.
- MSP2_INAV_MISC reports the active link (activeRX byte).
- rx_rx_unittest rewritten for the dual-link API (primary->secondary
  failover, secondary-active start, main-failsafe only when both down).

Regenerated docs/Settings.md and the MSP enum/message docs.

+++ src/main/fc/multifunction.h
@@ -28,7 +28,7 @@
 #include "drivers/display.h"
 
 typedef struct multiFunctionWarning_s {
-    uint8_t osdWarningsFlags;  // bitfield
+    uint16_t osdWarningsFlags;  // bitfield
     bool newWarningActive;
 } multiFunctionWarning_t;
